Move Layer-specific utilities to a separate file
