+++ src/types/network.ts
@@ -1,10 +1,5 @@
 import { TFunction } from 'i18next'
 
-// Here we need to import from the generated code, in order to break
-// a cycle of imports which confuse jest
-// eslint-disable-next-line no-restricted-imports
-import { Layer } from '../oasis-indexer/generated/api'
-
 export type Network = (typeof Network)[keyof typeof Network]
 
 // eslint-disable-next-line @typescript-eslint/no-redeclare
@@ -17,19 +12,6 @@ export const getNetworkNames = (t: TFunction): Record<Network, string> => ({
   [Network.mainnet]: t('common.mainnet'),
   [Network.testnet]: t('common.testnet'),
 })
-
-export const getLayerNames = (t: TFunction): Record<Layer, string> => ({
-  [Layer.emerald]: t('common.emerald'),
-  [Layer.sapphire]: t('common.sapphire'),
-  [Layer.cipher]: t('common.cipher'),
-  [Layer.consensus]: t('common.consensus'),
-})
-
-interface HasLayer {
-  layer: Layer
-}
-
-export const getFilterForLayer = (layer: Layer) => (item: HasLayer) => item.layer === layer
 
 interface HasNetwork {
   network: Network
